Added HardwareInfo as the main structure in the robot's URDF file
